TIP-1391: extract diff checker
